Isolate EIP712 signer to separate wasm entrypoint
