+++ packages/signers/src/index.ts
@@ -3,4 +3,3 @@ export * from "./snap";
 export * from "./ed25519";
 export * from "./secp256k1";
 export * from "./privy";
-export * from "./eip712";
