+++ packages/signers/src/wasm.ts
@@ -1,0 +1,2 @@
+export * from "./index";
+export * from "./wasm/eip712";
